Move heatmap code into separate source file
