+++ gmaps/__init__.py
@@ -2,6 +2,7 @@ from ._version import version_info, __version__  # noqa
 
 from .maps import *  # noqa
 from .marker import *  # noqa
+from .heatmap import *  # noqa
 
 
 def _jupyter_nbextension_paths():
